feat: change folder structure
