+++ apps/ui/src/app/[locale]/dev/pages-overview/components/PageList.tsx
@@ -3,7 +3,7 @@
 import { useState } from "react"
 
 import AppLink from "@/components/elementary/AppLink"
-import Typography from "@/components/html/components/Typography"
+import Typography from "@/components/typography"
 import { Button } from "@/components/ui/button"
 
 export default function PageList({
