+++ apps/ui/src/components/typography/index.tsx
@@ -1,32 +1,32 @@
 import type React from "react"
 
 import type {
-  ElementTag,
-  FontWeight,
   TextColor,
+  TypographyTag,
+  FontWeight,
   Variant,
-} from "@/components/html/config"
-import { resolveHtmlComponentClassName } from "@/components/html/helpers"
+} from "@/components/typography/config"
+import { resolveHtmlComponentClassName } from "@/components/typography/helpers"
 
-interface ElementProps {
+interface TypographyProps {
   readonly children: React.ReactNode
   readonly className?: string
   readonly variant?: Variant
   readonly textColor?: TextColor
   readonly fontWeight?: FontWeight
-  readonly tag: ElementTag
+  readonly tag?: TypographyTag
   readonly id?: string
 }
 
-export function Element({
+export function Typography({
   children,
   className,
   variant,
   textColor = "black",
   fontWeight = "normal",
-  tag: Tag,
+  tag: Tag = "p",
   id,
-}: ElementProps) {
+}: TypographyProps) {
   return (
     <Tag
       id={id}
@@ -43,6 +43,6 @@ export function Element({
   )
 }
 
-Element.displayName = "Element"
+Typography.displayName = "Typography"
 
-export default Element
+export default Typography
